uart: use 8250 omap serial driver instead of the omap serial driver
This fixes an issue that appears when transferring big amount of data
over bluetooth. Together with a not conform sending ble tag, this could
lead to a complete freeze of bluetooth communication. I don't understand
the exact problem, but it seems that the rx path freezes for some reasons.
On some later commits the most work on omap serial is related to dma
transfers, perhaps something is broken there. The CPU load doesn't seem
to become too high.